refactor(boolpolicy): merge boolexpr parser into boolpolicy package
Move parser.go and parser_test.go from token/core/common/boolexpr/ into
token/services/identity/boolpolicy/, eliminating the cross-package
dependency. Update sig.go, deserializer.go, and identity_test.go to use
the local Node/Parse symbols. Delete the now-empty boolexpr directory.
