+++ token/services/identity/boolpolicy/parser_test.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package boolexpr
+package boolpolicy
 
 import (
 	"testing"
